+++ pkg/analysis_server/test/src/services/correction/fix/remove_comparison_test.dart
@@ -281,7 +281,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) =>
+  operator ==(Object other) =>
           other != null &&
           other is Person &&
           name == other.name;
@@ -292,7 +292,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) =>
+  operator ==(Object other) =>
           other is Person &&
           name == other.name;
 }
@@ -305,7 +305,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     return other != null &&
           other is Person &&
           name == other.name;
@@ -317,7 +317,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     return other is Person &&
           name == other.name;
   }
@@ -331,7 +331,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     if (other is! Person) return false;
     other ??= Person();
     return other.name == name;
@@ -349,7 +349,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     if (other is! Person) return false;
     final toCompare = other ?? Person();
     return toCompare.name == name;
@@ -362,7 +362,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     if (other is! Person) return false;
     final toCompare = other;
     return toCompare.name == name;
@@ -381,7 +381,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     if (other == null) return false;
     return other is Person &&
           name == other.name;
@@ -393,7 +393,7 @@ class Person {
   final String name = '';
 
   @override
-  operator ==(Object? other) {
+  operator ==(Object other) {
     return other is Person &&
           name == other.name;
   }