+++ runtime/vm/compiler/jit/compiler.cc
@@ -313,7 +313,6 @@ class CompileParsedFunctionHelper : public ValueObject {
   bool optimized() const { return optimized_; }
   intptr_t osr_id() const { return osr_id_; }
   Thread* thread() const { return thread_; }
-  Isolate* isolate() const { return thread_->isolate(); }
   IsolateGroup* isolate_group() const { return thread_->isolate_group(); }
   CodePtr FinalizeCompilation(compiler::Assembler* assembler,
                               FlowGraphCompiler* graph_compiler,
@@ -879,7 +878,9 @@ ErrorPtr Compiler::EnsureUnoptimizedCode(Thread* thread,
 ObjectPtr Compiler::CompileOptimizedFunction(Thread* thread,
                                              const Function& function,
                                              intptr_t osr_id) {
-  VMTagScope tagScope(thread, VMTag::kCompileOptimizedTagId);
+  VMTagScope tag_scope(thread, VMTag::kCompileOptimizedTagId);
+  NoActiveIsolateScope no_active_isolate(thread);
+
 #if defined(SUPPORT_TIMELINE)
   const char* event_name;
   if (osr_id != kNoOSRDeoptId) {